+++ public/sw.js
@@ -1,4 +1,4 @@
-const SHELL_CACHE = 'darkroom-v115';
+const SHELL_CACHE = 'darkroom-v116';
 const THUMB_CACHE = 'darkroom-thumbs-v1';
 const THUMB_MAX = 500;